getting somewhere
Play with a few algorithms for picking random stations.
Trying to prefer gentle changes in the path seems to work best so far.
When picking a station, pick one that gives the smallest change in slope
when compared with the slope between the current station and previous
station.
Add some logging, and fix color selection in this testing modeso routes
don't duplicate colors.